centarl authState
НавБар убран в компоненты, стейт на статус авторизации вынесен в App. Проведена подготовка к MUI
